+++ SeukCalendar/TodayScheduleWidget/Components/WidgetCalendarWeekdayHeader/WidgetCalendarWeekdayHeader.swift
@@ -4,6 +4,8 @@ import SwiftUI
 import WidgetKit
 
 struct WidgetCalendarWeekdayHeader: View {
+  @Environment(\.widgetRenderingMode) private var widgetRenderingMode
+
   private enum Metrics {
     static let weekdayFont = Widget.Large.medium
     static let weekdayHeight: CGFloat = 24
@@ -22,7 +24,12 @@ struct WidgetCalendarWeekdayHeader: View {
         ForEach(Array(configuration.weekdayItems.enumerated()), id: \.offset) { _, item in
           Text(item.title)
             .font(Metrics.weekdayFont)
-            .foregroundStyle(item.role.foregroundColor)
+            .foregroundStyle(
+              configuration.resolvedForegroundColor(
+                for: item.role,
+                widgetRenderingMode: widgetRenderingMode
+              )
+            )
             .lineLimit(1)
             .minimumScaleFactor(0.7)
             .frame(height: Metrics.weekdayHeight)
@@ -41,65 +48,9 @@ struct WidgetCalendarWeekdayHeader: View {
 private extension WidgetCalendarWeekdayHeader {
   var divider: some View {
     Rectangle()
-      .fill(Color.semantic.Background.secondary)
+      .fill(configuration.resolvedDividerColor(for: widgetRenderingMode))
       .frame(maxWidth: .infinity)
       .frame(height: Metrics.dividerHeight)
-  }
-}
-
-extension WidgetCalendarWeekdayHeader {
-  struct Configuration: Hashable {
-    let locale: Locale
-
-    init(locale: Locale = Locale(identifier: "ko_KR")) {
-      self.locale = locale
-    }
-  }
-}
-
-private extension WidgetCalendarWeekdayHeader.Configuration {
-  struct WeekdayItem: Hashable {
-    enum Role: Hashable {
-      case sunday
-      case weekday
-      case saturday
-    }
-
-    let title: String
-    let role: Role
-  }
-
-  var weekdayItems: [WeekdayItem] {
-    var calendar = Calendar(identifier: .gregorian)
-    calendar.locale = locale
-    calendar.firstWeekday = 1
-
-    return calendar.shortStandaloneWeekdaySymbols.enumerated().map { index, title in
-      let role: WeekdayItem.Role
-      switch index {
-      case 0:
-        role = .sunday
-      case 6:
-        role = .saturday
-      default:
-        role = .weekday
-      }
-
-      return WeekdayItem(title: title, role: role)
-    }
-  }
-}
-
-private extension WidgetCalendarWeekdayHeader.Configuration.WeekdayItem.Role {
-  var foregroundColor: Color {
-    switch self {
-    case .sunday:
-      .semanticExtensions.Content.contentNegative
-    case .weekday:
-      .semantic.Content.secondary
-    case .saturday:
-      .semanticExtensions.Content.contentAccent
-    }
   }
 }
